docs/requirements.txt
sphinx==8.0.2
intersphinx-registry
myst-parser
numpydoc
pydata-sphinx-theme
